[#3036] use uint64_t for parameter and uint32_t as return, adjust test

+++ nes-runtime/src/Execution/Expressions/Functions/BitcounterExpression.cpp
@@ -21,7 +21,7 @@ namespace NES::Runtime::Execution::Expressions {
 BitcounterExpression::BitcounterExpression(const NES::Runtime::Execution::Expressions::ExpressionPtr& SubExpression)
     : SubExpression(SubExpression) {}
 
-int bitcounter(uint number) {
+uint32_t bitcounter(uint64_t number) {
     return std::popcount(number);
     }
